🚚 レビューの修正(analysisページ移動とコンポーネント分割)
