+++ app/(analysis)/_components/key-indicator.tsx
@@ -1,6 +1,4 @@
-import { Suspense } from "react"
-import type { Metadata } from "next"
-import { env } from "@/env.mjs"
+import { env } from "process"
 import { ArrowDownRight, ArrowUpRight } from "lucide-react"
 
 import { type Analysis } from "@/types/api/analysis"
@@ -9,66 +7,10 @@ import {
   fetcher,
   getAbsoluteValueRoundedToOneDecimal,
 } from "@/lib/utils"
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { Badge } from "@/components/ui/badge"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { ChannelStatsSkeleton } from "@/components/skeletons/channel-stats-skeleton"
-import { KeyIndicatorSkeleton } from "@/components/skeletons/key-indicator-skeleton"
 
-export const metadata: Metadata = {
-  title: "分析画面",
-  description: "Generated by create next app",
-}
-
-export default function AnalysisPage() {
-  // sessionから取得？
-  const token = "c7d091db-8b2b-4b62-a71f-0a13b6a42c1a"
-
-  return (
-    <div className="flex flex-col gap-10">
-      <Suspense fallback={<ChannelStatsSkeleton />}>
-        <ChannelStats token={token} />
-      </Suspense>
-      <div className="flex flex-col gap-2">
-        <div className="flex items-center justify-between">
-          <p className="text-2xl font-bold">重要指標</p>
-          <p className="text-sm text-muted-foreground">過去28日間</p>
-        </div>
-        <Suspense fallback={<KeyIndicatorSkeleton count={2} />}>
-          <KeyIndicator token={token} />
-        </Suspense>
-      </div>
-      <div>
-        <p className="text-2xl font-bold">コメント返信状況</p>
-      </div>
-    </div>
-  )
-}
-
-async function ChannelStats({ token }: { token: string }) {
-  const { user, analytics } = await fetcher<Analysis>({
-    url: `${env.API_ENDPOINT}/mock/analysis`,
-    headers: { token: token },
-  })
-
-  return (
-    <div className="flex items-center gap-3 self-stretch">
-      <Avatar className="h-20 w-20">
-        <AvatarImage src={user.image} />
-        <AvatarFallback>{user.name.slice(0, 2).toUpperCase()}</AvatarFallback>
-      </Avatar>
-      <div className="flex flex-col items-start justify-center">
-        <p className="line-clamp-1 text-lg">{user.name}</p>
-        <p className="text-2xl font-bold">
-          {analytics.subscribers.toLocaleString()}
-        </p>
-        <p className="text-sm text-muted-foreground">チャンネル登録者</p>
-      </div>
-    </div>
-  )
-}
-
-async function KeyIndicator({ token }: { token: string }) {
+export async function KeyIndicator({ token }: { token: string }) {
   const { analytics } = await fetcher<Analysis>({
     url: `${env.API_ENDPOINT}/mock/analysis`,
     headers: { token: token },
