+++ app/(analysis)/page.tsx
@@ -1,0 +1,38 @@
+import { Suspense } from "react"
+import type { Metadata } from "next"
+
+import { ChannelStatsSkeleton } from "@/components/skeletons/channel-stats-skeleton"
+import { KeyIndicatorSkeleton } from "@/components/skeletons/key-indicator-skeleton"
+
+import { ChannelStats } from "./_components/channel-stats"
+import { KeyIndicator } from "./_components/key-indicator"
+
+export const metadata: Metadata = {
+  title: "分析画面",
+  description: "Generated by create next app",
+}
+
+export default function AnalysisPage() {
+  // sessionから取得？
+  const token = "c7d091db-8b2b-4b62-a71f-0a13b6a42c1a"
+
+  return (
+    <div className="flex flex-col gap-10">
+      <Suspense fallback={<ChannelStatsSkeleton />}>
+        <ChannelStats token={token} />
+      </Suspense>
+      <div className="flex flex-col gap-2">
+        <div className="flex items-center justify-between">
+          <p className="text-2xl font-bold">重要指標</p>
+          <p className="text-sm text-muted-foreground">過去28日間</p>
+        </div>
+        <Suspense fallback={<KeyIndicatorSkeleton count={2} />}>
+          <KeyIndicator token={token} />
+        </Suspense>
+      </div>
+      <div>
+        <p className="text-2xl font-bold">コメント返信状況</p>
+      </div>
+    </div>
+  )
+}
